# app/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "app, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Users": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json;odata.metadata=minimal;odata.streaming=true",
          "application/json;odata.metadata=minimal;odata.streaming=false",
          "application/json;odata.metadata=minimal",
          "application/json;odata.metadata=full;odata.streaming=true",
          "application/json;odata.metadata=full;odata.streaming=false",
          "application/json;odata.metadata=full",
          "application/json;odata.metadata=none;odata.streaming=true",
          "application/json;odata.metadata=none;odata.streaming=false",
          "application/json;odata.metadata=none",
          "application/json;odata.streaming=true",
          "application/json;odata.streaming=false",
          "application/json",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=full;IEEE754Compatible=false",
          "application/json;odata.metadata=full;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=true",
          "application/json;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.streaming=false;IEEE754Compatible=true",
          "application/json;IEEE754Compatible=false",
          "application/json;IEEE754Compatible=true",
          "application/xml",
          "text/plain",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json;odata.metadata=minimal;odata.streaming=true",
          "application/json;odata.metadata=minimal;odata.streaming=false",
          "application/json;odata.metadata=minimal",
          "application/json;odata.metadata=full;odata.streaming=true",
          "application/json;odata.metadata=full;odata.streaming=false",
          "application/json;odata.metadata=full",
          "application/json;odata.metadata=none;odata.streaming=true",
          "application/json;odata.metadata=none;odata.streaming=false",
          "application/json;odata.metadata=none",
          "application/json;odata.streaming=true",
          "application/json;odata.streaming=false",
          "application/json",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=full;IEEE754Compatible=false",
          "application/json;odata.metadata=full;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=true",
          "application/json;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.streaming=false;IEEE754Compatible=true",
          "application/json;IEEE754Compatible=false",
          "application/json;IEEE754Compatible=true",
          "application/xml",
          "text/plain",
          "application/octet-stream",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/Users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "produces": [
          "application/json;odata.metadata=minimal;odata.streaming=true",
          "application/json;odata.metadata=minimal;odata.streaming=false",
          "application/json;odata.metadata=minimal",
          "application/json;odata.metadata=full;odata.streaming=true",
          "application/json;odata.metadata=full;odata.streaming=false",
          "application/json;odata.metadata=full",
          "application/json;odata.metadata=none;odata.streaming=true",
          "application/json;odata.metadata=none;odata.streaming=false",
          "application/json;odata.metadata=none",
          "application/json;odata.streaming=true",
          "application/json;odata.streaming=false",
          "application/json",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=full;IEEE754Compatible=false",
          "application/json;odata.metadata=full;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=true",
          "application/json;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.streaming=false;IEEE754Compatible=true",
          "application/json;IEEE754Compatible=false",
          "application/json;IEEE754Compatible=true",
          "application/xml",
          "text/plain",
          "application/octet-stream",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json;odata.metadata=minimal;odata.streaming=true",
          "application/json;odata.metadata=minimal;odata.streaming=false",
          "application/json;odata.metadata=minimal",
          "application/json;odata.metadata=full;odata.streaming=true",
          "application/json;odata.metadata=full;odata.streaming=false",
          "application/json;odata.metadata=full",
          "application/json;odata.metadata=none;odata.streaming=true",
          "application/json;odata.metadata=none;odata.streaming=false",
          "application/json;odata.metadata=none",
          "application/json;odata.streaming=true",
          "application/json;odata.streaming=false",
          "application/json",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=minimal;IEEE754Compatible=false",
          "application/json;odata.metadata=minimal;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=full;IEEE754Compatible=false",
          "application/json;odata.metadata=full;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true",
          "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=false",
          "application/json;odata.metadata=none;IEEE754Compatible=true",
          "application/json;odata.streaming=true;IEEE754Compatible=false",
          "application/json;odata.streaming=true;IEEE754Compatible=true",
          "application/json;odata.streaming=false;IEEE754Compatible=false",
          "application/json;odata.streaming=false;IEEE754Compatible=true",
          "application/json;IEEE754Compatible=false",
          "application/json;IEEE754Compatible=true",
          "application/xml",
          "text/plain",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Category": {
      "type": "object",
      "properties": {
        "categoryId": {
          "format": "int32",
          "type": "integer"
        },
        "categoryName": {
          "type": "string"
        },
        "stories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Story"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Chapter": {
      "type": "object",
      "properties": {
        "chapterId": {
          "format": "int64",
          "type": "integer"
        },
        "chapterNumber": {
          "format": "int64",
          "type": "integer"
        },
        "storyId": {
          "format": "int32",
          "type": "integer"
        },
        "volumeId": {
          "format": "int32",
          "type": "integer"
        },
        "chapterTitle": {
          "type": "string"
        },
        "chapterContent": {
          "type": "string"
        },
        "chapterPrice": {
          "format": "double",
          "type": "number"
        },
        "createTime": {
          "format": "date-time",
          "type": "string"
        },
        "updateTime": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "reportContents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReportContent"
          }
        },
        "story": {
          "$ref": "#/definitions/Story"
        },
        "storyReads": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StoryRead"
          }
        },
        "volume": {
          "$ref": "#/definitions/Volume"
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      },
      "additionalProperties": false
    },
    "Comment": {
      "type": "object",
      "properties": {
        "commentId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "storyId": {
          "format": "int32",
          "type": "integer"
        },
        "chapterId": {
          "format": "int64",
          "type": "integer"
        },
        "commentContent": {
          "type": "string"
        },
        "commentDate": {
          "format": "date-time",
          "type": "string"
        },
        "chapter": {
          "$ref": "#/definitions/Chapter"
        },
        "reportContents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReportContent"
          }
        },
        "story": {
          "$ref": "#/definitions/Story"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "additionalProperties": false
    },
...
    "ReportContent": {
      "type": "object",
      "properties": {
        "reportId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "reportTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "storyId": {
          "format": "int32",
          "type": "integer"
        },
        "chapterId": {
          "format": "int64",
          "type": "integer"
        },
        "commentId": {
          "format": "int32",
          "type": "integer"
        },
        "reportContent1": {
          "type": "string"
        },
        "reportDate": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "type": "boolean"
        },
        "chapter": {
          "$ref": "#/definitions/Chapter"
        },
        "comment": {
          "$ref": "#/definitions/Comment"
        },
        "reportType": {
          "$ref": "#/definitions/ReportType"
        },
        "story": {
          "$ref": "#/definitions/Story"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "additionalProperties": false
    },
...
    "ReportType": {
      "type": "object",
      "properties": {
        "reportTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "reportTypeContent": {
          "type": "string"
        },
        "reportContents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReportContent"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Story": {
      "type": "object",
      "properties": {
        "storyId": {
          "format": "int32",
          "type": "integer"
        },
        "storyTitle": {
          "type": "string"
        },
        "authorId": {
          "format": "int32",
          "type": "integer"
        },
        "storyPrice": {
          "format": "double",
          "type": "number"
        },
        "storySale": {
          "format": "double",
          "type": "number"
        },
        "storyImage": {
          "type": "string"
        },
        "storyDescription": {
          "type": "string"
        },
        "createTime": {
          "format": "date-time",
          "type": "string"
        },
        "updateTime": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "author": {
          "$ref": "#/definitions/User"
        },
        "chapters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Chapter"
          }
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "reportContents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReportContent"
          }
        },
        "storyFollowLikes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StoryFollowLike"
          }
        },
        "storyInteraction": {
          "$ref": "#/definitions/StoryInteraction"
        },
        "storyReads": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StoryRead"
          }
        },
        "volumes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Volume"
          }
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      },
      "additionalProperties": false
    },
    "StoryFollowLike": {
      "type": "object",
      "properties": {
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "storyId": {
          "format": "int32",
          "type": "integer"
        },
        "follow": {
          "type": "boolean"
        },
        "like": {
          "type": "boolean"
        },
        "story": {
          "$ref": "#/definitions/Story"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "additionalProperties": false
    },
    "StoryInteraction": {
      "type": "object",
      "properties": {
        "storyId": {
          "format": "int32",
          "type": "integer"
        },
        "like": {
          "format": "int32",
          "type": "integer"
        },
        "follow": {
          "format": "int32",
          "type": "integer"
        },
        "view": {
          "format": "int32",
          "type": "integer"
        },
        "read": {
          "format": "int32",
          "type": "integer"
        },
        "story": {
          "$ref": "#/definitions/Story"
        }
      },
      "additionalProperties": false
    },
    "StoryRead": {
      "type": "object",
      "properties": {
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "storyId": {
          "format": "int32",
          "type": "integer"
        },
        "chapterId": {
          "format": "int64",
          "type": "integer"
        },
        "readTime": {
          "format": "date-time",
          "type": "string"
        },
        "chapter": {
          "$ref": "#/definitions/Chapter"
        },
        "story": {
          "$ref": "#/definitions/Story"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "additionalProperties": false
    },
    "Transaction": {
      "type": "object",
      "properties": {
        "transactionId": {
          "format": "int64",
          "type": "integer"
        },
        "walletId": {
          "format": "int32",
          "type": "integer"
        },
        "amount": {
          "format": "double",
          "type": "number"
        },
        "fundBefore": {
          "format": "double",
          "type": "number"
        },
        "fundAfter": {
          "format": "double",
          "type": "number"
        },
        "refundBefore": {
          "format": "double",
          "type": "number"
        },
        "refundAfter": {
          "format": "double",
          "type": "number"
        },
        "transactionTime": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "type": "boolean"
        },
        "description": {
          "type": "string"
        },
        "wallet": {
          "$ref": "#/definitions/Wallet"
        }
      },
      "additionalProperties": false
    },
    "User": {
      "type": "object",
      "properties": {
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "userFullname": {
          "type": "string"
        },
        "gender": {
          "type": "boolean"
        },
        "dob": {
          "format": "date-time",
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "userImage": {
          "type": "string"
        },
        "status": {
          "type": "boolean"
        },
        "descriptionMarkdown": {
          "type": "string"
        },
        "descriptionHtml": {
          "type": "string"
        },
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        },
        "reportContents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReportContent"
          }
        },
        "stories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Story"
          }
        },
        "storyFollowLikes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StoryFollowLike"
          }
        },
        "storyReads": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StoryRead"
          }
        },
        "wallets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Wallet"
          }
        },
        "chapters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Chapter"
          }
        },
        "storiesNavigation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Story"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Volume": {
      "type": "object",
      "properties": {
        "volumeId": {
          "format": "int32",
          "type": "integer"
        },
        "storyId": {
          "format": "int32",
          "type": "integer"
        },
        "volumeTitle": {
          "type": "string"
        },
        "chapters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Chapter"
          }
        },
        "story": {
          "$ref": "#/definitions/Story"
        }
      },
      "additionalProperties": false
    },
    "Wallet": {
      "type": "object",
      "properties": {
        "walletId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "fund": {
          "format": "double",
          "type": "number"
        },
        "refund": {
          "format": "double",
          "type": "number"
        },
        "transactions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Transaction"
          }
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "additionalProperties": false
    }
  }
}